예외 코드 추가_1

+++ src/main/java/com/example/demo/service/BoardService.java
@@ -2,6 +2,9 @@ package com.example.demo.service;
 
 import java.util.List;
 
+import com.example.demo.Exception.ArticleExist;
+import com.example.demo.Exception.BoardNotFound;
+import com.example.demo.repository.ArticleRepositoryJdbc;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -16,9 +19,11 @@ import com.example.demo.repository.BoardRepository;
 public class BoardService {
 
     private final BoardRepository boardRepository;
+    private final ArticleRepositoryJdbc articleRepositoryJdbc;
 
-    public BoardService(BoardRepository boardRepository) {
+    public BoardService(BoardRepository boardRepository, ArticleRepositoryJdbc articleRepositoryJdbc) {
         this.boardRepository = boardRepository;
+        this.articleRepositoryJdbc = articleRepositoryJdbc;
     }
 
     public List<BoardResponse> getBoards() {
@@ -29,6 +34,9 @@ public class BoardService {
 
     public BoardResponse getBoardById(Long id) {
         Board board = boardRepository.findById(id);
+        if (board == null) {
+            throw new BoardNotFound("게시판을 찾을 수 없습니다.");
+        }
         return BoardResponse.from(board);
     }
 
@@ -41,12 +49,17 @@ public class BoardService {
 
     @Transactional
     public void deleteBoard(Long id) {
+        if (articleRepositoryJdbc.existByBoardId(id))
+            throw new ArticleExist("게시물이 존재합니다.");
         boardRepository.deleteById(id);
     }
 
     @Transactional
     public BoardResponse update(Long id, BoardUpdateRequest request) {
         Board board = boardRepository.findById(id);
+        if (board == null) {
+            throw new BoardNotFound("게시판을 찾을 수 없습니다.");
+        }
         board.update(request.name());
         Board updated = boardRepository.update(board);
         return BoardResponse.from(updated);
